New: Extract  stylish formatter
